+++ telnet/libtelnet/auth.c
@@ -382,7 +382,7 @@ auth_send(data, cnt)
 	 * Save the list of authentication mechanisms
 	 */
 	auth_send_cnt = cnt;
-	if (auth_send_cnt > sizeof(_auth_send_data))
+	if ((size_t) auth_send_cnt > sizeof(_auth_send_data))
 	    auth_send_cnt = sizeof(_auth_send_data);
 	memcpy(_auth_send_data, data, (unsigned) auth_send_cnt);
 	auth_send_data = _auth_send_data;
@@ -504,7 +504,7 @@ auth_name(data, cnt)
 			printf(">>>%s: Empty name in NAME\r\n", Name);
 		return;
 	}
-	if (cnt > sizeof(savename) - 1) {
+	if ((size_t) cnt > sizeof(savename) - 1) {
 		if (auth_debug_mode)
 			printf(">>>%s: Name in NAME (%d) exceeds %d length\r\n",
 					Name, cnt, (int) sizeof(savename)-1);
